+++ R/rats.R
@@ -217,11 +217,23 @@ optim_rats <- function(x, ref, t, inds, basis, max_shift) {
   
   if (is.null(basis)) {
     basis_mod <- x
-    ahat <- unname(qr.solve(basis_mod, ref))
+    ahat <- try(unname(qr.solve(basis_mod, ref)))
+    
+    # use ginv if qr.solve fails
+    if (class(ahat) == "try-error") {
+      ahat <- unname(ginv(basis_mod) %*% ref)
+      warning("RATS qr.solve failed, reverted to ginv, check for poor data")
+    }
     yhat <- basis_mod * ahat
   } else {
     basis_mod <- cbind(x, basis)
-    ahat <- unname(qr.solve(basis_mod, ref))
+    ahat <- try(unname(qr.solve(basis_mod, ref)))
+    
+    # use ginv if qr.solve fails
+    if (class(ahat) == "try-error") {
+      ahat <- unname(ginv(basis_mod) %*% ref)
+      warning("RATS qr.solve failed, reverted to ginv, check for poor data")
+    }
     yhat <- basis_mod %*% ahat
     bl   <- basis_mod %*% c(0, ahat[2:length(ahat)])
   }
@@ -242,12 +254,11 @@ rats_obj_fn <- function(par, x, ref, t, inds, basis) {
     basis_mod <- cbind(x, basis)
   }
   
-  # use ginv
-  #inv_basis <- ginv(basis)
-  #ahat <- inv_basis %*% ref
+  # try use qr as it's faster
+  ahat <- try(qr.solve(basis_mod, ref), silent = TRUE)
   
-  # use qr
-  ahat <- qr.solve(basis_mod, ref)
+  # use ginv if qr.solve fails
+  if (class(ahat) == "try-error") ahat <- ginv(basis_mod) %*% ref
   
   if (is.null(basis)) {
     yhat <- basis_mod * ahat